+++ src/js/browser-action/tippy-footer.js
@@ -1,0 +1,8 @@
+export default function() {
+  tippy('#footer [data-tippy-content]', {
+    theme: 'light-border',
+    arrow: false,
+    trigger: 'mouseenter',
+    offset: [0, 5]
+  });
+}