+++ src/js/grid-page.js
@@ -2,6 +2,7 @@ import drawGrid        from './browser-action/draw-grid.js';
 import drawCommandHelp from './browser-action/draw-command-help.js';
 import includeHTML     from './browser-action/include-html.js';
 import tippyHelp       from './browser-action/tippy-help.js';
+import tippyFooter     from './browser-action/tippy-footer.js';
 
 chrome.runtime.getBackgroundPage((page) => {
   const app = page.app;
@@ -10,6 +11,7 @@ chrome.runtime.getBackgroundPage((page) => {
       drawGrid(page);
       commands.forEach(drawCommandHelp);
       feather.replace();
+      tippyFooter();
       tippyHelp();
     });
   });